+++ week2_assignment/css/style.css
@@ -1,5 +1,5 @@
 /* reset css */
-@import url('./reset.css');
+@import url('../css/reset.css');
 
 html{
   scroll-behavior: smooth;
