+++ week2_assignment/html/cart.html
@@ -4,13 +4,13 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <script src="https://kit.fontawesome.com/455c79571e.js" crossorigin="anonymous"></script>
-    <link rel="stylesheet" href="./css/cart.css">
+    <link rel="stylesheet" href="../css/cart.css">
     <title>wishlist_cart</title>
 </head>
 <body>
   <!-- cart 헤더 -->
     <header class="cart-header-section">
-        <a href="home.html"><i class="fa-solid fa-house"></i></a>
+        <a href="../html/home.html"><i class="fa-solid fa-house"></i></a>
         <h1 class="header__title">장바구니 페이지</h1>
         <i class="fa-solid fa-bars"></i>
     </header>
@@ -36,7 +36,7 @@
       </table>
       </section>
       <button id="buy-btn">구매하기</button>
-      <a href="./home.html"><button id="home-btn">홈으로</button></a>
+      <a href="../html/home.html"><button id="home-btn">홈으로</button></a>
   </main>
 <!-- 모달 -->
   <div class="modal" style="display: none;">
@@ -49,7 +49,7 @@
     </div>
 </div>
 
-  <script src="./cart.js"></script>
+  <script src="../js/cart.js"></script>
 
 </body>
 </html>
