+++ week2_assignment/html/home.html
@@ -5,12 +5,12 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>건휘's market</title>
     <script src="https://kit.fontawesome.com/455c79571e.js" crossorigin="anonymous"></script>
-    <link rel="stylesheet" href="./css/style.css">
+    <link rel="stylesheet" href="../css/style.css">
 </head>
     <body>
         <!-- 헤더 -->
         <header class="header-section">
-            <a href="home.html"><i class="fa-solid fa-house"></i></a>
+            <a href="../html/home.html"><i class="fa-solid fa-house"></i></a>
             <h1 class="header__title">건휘's favorite</h1>
             <i class="fa-solid fa-bars" onclick="openNav()"></i>
         </header>
@@ -21,7 +21,7 @@
                 <p class="back-btn" onclick="closeNav()">&#10132;</p>
                 <a>관심 상품 목록</a>
                 <a>관리자 페이지</a>
-                <a href="cart.html">장바구니</a>
+                <a href="../html/cart.html">장바구니</a>
             </ul>
         </nav>
         <!-- 카테고리 -->
@@ -51,7 +51,7 @@
             </ul>
         </footer>
     </body>
-    <script src="./sideNav.js"></script>
-    <script src="./main.js" type="module"></script>
-    <script src="./addCart.js"></script>
+    <script src="../js/sideNav.js"></script>
+    <script src="../js/main.js" type="module"></script>
+    <script src="../js/addCart.js"></script>
 </html>
