+++ week2_assignment/js/const.js
@@ -5,105 +5,105 @@ export const SHOPPING_LIST = [
         name: "딸기",
         category: "food",
         price: 12000,
-        img: "./img/food/berry.jpg", 
+        img: "../img/food/berry.jpg", 
     },
     {
         id: 2,
         name: "치킨",
         category: "food",
         price: 20000,
-        img: "./img/food/chicken.jpg", 
+        img: "../img/food/chicken.jpg", 
     },
     {
         id: 3,
         name: "막창",
         category: "food",
         price: 12000,
-        img: "./img/food/makchang.jpg", 
+        img: "../img/food/makchang.jpg", 
     },
     {
         id: 4,
         name: "마라탕",
         category: "food",
         price: 12000,
-        img: "./img/food/maratang.jpg", 
+        img: "../img/food/maratang.jpg", 
     },
     {
         id: 5,
         name: "샤인머스켓",
         category: "food",
         price: 9900,
-        img: "./img/food/musket.jpg", 
+        img: "../img/food/musket.jpg", 
     },
     {
         id: 6,
         name: "아스날",
         category: "sports",
         price: 100000,
-        img: "./img/sports/arsenal.jpg", 
+        img: "../img/sports/arsenal.jpg", 
     },
     {
         id: 7,
         name: "자이언츠",
         category: "sports",
         price: 80000,
-        img: "./img/sports/giants.jpg", 
+        img: "../img/sports/giants.jpg", 
     },
     {
         id: 8,
         name: "골든스테이트",
         category: "sports",
         price: 60000,
-        img: "./img/sports/golden.jpg", 
+        img: "../img/sports/golden.jpg", 
     },
     {
         id: 9,
         name: "기아",
         category: "sports",
         price: 18000,
-        img: "./img/sports/kia.jpg", 
+        img: "../img/sports/kia.jpg", 
     },
     {
         id: 10,
         name: "t1",
         category: "sports",
         price: 30000,
-        img: "./img/sports/t1.jpg", 
+        img: "../img/sports/t1.jpg", 
     },
     {
         id: 11,
         name: "독서",
         category: "hobby",
         price: 12000,
-        img: "./img/hobby/book.jpg", 
+        img: "../img/hobby/book.jpg", 
     },
     {
         id: 12,
         name: "바둑",
         category: "hobby",
         price: 5000,
-        img: "./img/hobby/go.jpg", 
+        img: "../img/hobby/go.jpg", 
     },
     {
         id: 13,
         name: "음악듣기",
         category: "hobby",
         price: 5000,
-        img: "./img/hobby/music.jpg", 
+        img: "../img/hobby/music.jpg", 
     },
     {
         id: 14,
         name: "산책",
         category: "hobby",
         price: 20000,
-        img: "./img/hobby/sancheck.jpg", 
+        img: "../img/hobby/sancheck.jpg", 
     },
     {
         id: 15,
         name: "잠",
         category: "hobby",
         price: 10000,
-        img: "./img/hobby/sleep.jpg", 
+        img: "../img/hobby/sleep.jpg", 
     },
 ]
